refactor: product 도메인 안에서 분리되는 컴포넌트이기에 product 디렉토리 안에 모든 도메인을 집중시킴
