+++ app/product/cart/components/CartItemList.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import CartProductCard from "@/app/cart/components/CartItem";
-import useCartStore from "@/app/cart/stores/useCartStore";
+import CartItem from "@/app/product/cart/components/CartItem";
+import useCartStore from "@/app/product/cart/stores/useCartStore";
 
 type CartItemListProps = {
   onProductRemove?: (itemId: string) => void;
@@ -18,7 +18,7 @@ function CartItemList({ onProductRemove }: CartItemListProps) {
   return (
     <div className="space-y-4">
       {items.map((item) => (
-        <CartProductCard key={item.id} item={item} onRemove={handleRemove} />
+        <CartItem key={item.id} item={item} onRemove={handleRemove} />
       ))}
     </div>
   );
